django_project/drf_pro/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.2
importlib-metadata==8.0.0
Markdown==3.6
sqlparse==0.5.0
typing-extensions==4.12.2
zipp==3.19.2
